refactor(search): Moves search form into a standalone component
